+++ app/src/main/java/pl/mattiahit/androidweatherk/di/components/AppComponent.kt
@@ -4,8 +4,8 @@ import dagger.Component
 import pl.mattiahit.androidweatherk.MainActivity
 import pl.mattiahit.androidweatherk.di.modules.*
 import pl.mattiahit.androidweatherk.home.ui.HomeFragment
-import pl.mattiahit.androidweatherk.home.domain.repository.LocationRepository
-import pl.mattiahit.androidweatherk.home.domain.repository.WeatherRepository
+import pl.mattiahit.androidweatherk.home.data.repository.LocationRepository
+import pl.mattiahit.androidweatherk.home.data.repository.WeatherRepository
 import pl.mattiahit.androidweatherk.home.ui.viewmodel.HomeViewModel
 import javax.inject.Singleton
