+++ app/src/main/java/pl/mattiahit/androidweatherk/di/modules/LiveDataModule.kt
@@ -3,7 +3,7 @@ package pl.mattiahit.androidweatherk.di.modules
 import dagger.Module
 import dagger.Provides
 import pl.mattiahit.androidweatherk.WeatherApplication
-import pl.mattiahit.androidweatherk.home.domain.livedata.LocationLiveData
+import pl.mattiahit.androidweatherk.home.data.livedata.LocationLiveData
 import javax.inject.Singleton
 
 @Module
