+++ app/src/main/res/drawable-night/clear.png
