+++ app/src/main/res/drawable-night/cloudy.png
