+++ app/src/main/res/drawable-night/stormy.png
