+++ app/src/main/res/drawable/clear.png
