+++ app/src/main/res/layout-land/fragment_home_v2.xml
@@ -36,7 +36,7 @@
             android:layout_centerVertical="true"
             android:layout_marginStart="5dp"
             android:scaleType="fitCenter"
-            android:src="@drawable/cloudy_day"
+            android:src="@drawable/cloudy"
             tools:ignore="ContentDescription" />
 
         <TextView
